Added 'can' method to client, moved namespace
